# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties": {
      "get": {
        "operationId": "PropertyController_findApproved",
        "summary": "List approved properties (public catalog)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "purpose",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "SALE",
                "RENT"
              ],
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Subcategory ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Cairo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      },
      "post": {
        "operationId": "PropertyController_create",
        "summary": "Create property as DRAFT (verified owner)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created"
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/my/list": {
      "get": {
        "operationId": "PropertyController_findMine",
        "summary": "List my properties (owner)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SOLD",
                "RENTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/{id}": {
      "get": {
        "operationId": "PropertyController_findOne",
        "summary": "Get property details",
        "description": "Public: APPROVED only. With Bearer token: owner sees own property (any status), admin sees all.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "uuid-here",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      },
      "patch": {
        "operationId": "PropertyController_update",
        "summary": "Update property (DRAFT or REJECTED only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "PropertyController_remove",
        "summary": "Delete property (DRAFT or REJECTED only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/{id}/images": {
      "post": {
        "operationId": "PropertyController_uploadImages",
        "summary": "Upload property images",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadPropertyImagesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/{id}/images/{imageId}": {
      "patch": {
        "operationId": "PropertyController_updateImage",
        "summary": "Update image order or primary flag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "PropertyController_removeImage",
        "summary": "Delete property image",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/{id}/submit": {
      "post": {
        "operationId": "PropertyController_submit",
        "summary": "Submit property for admin review (→ PENDING)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/{id}/mark-sold": {
      "patch": {
        "operationId": "PropertyController_markSold",
        "summary": "Mark approved SALE property as SOLD",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/properties/{id}/mark-rented": {
      "patch": {
        "operationId": "PropertyController_markRented",
        "summary": "Mark approved RENT property as RENTED",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Aqar API",
    "description": "Rental Property Platform — Authentication, Authorization (RBAC), Properties & Bookings",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Modern 3BR Apartment in Nasr City"
          },
          "description": {
            "type": "string",
            "example": "Spacious apartment with balcony and parking."
          },
          "price": {
            "type": "number",
            "example": 2500000
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "area": {
            "type": "string",
            "example": "Nasr City"
          },
          "address": {
            "type": "string",
            "example": "123 Abbas El Akkad Street"
          },
          "latitude": {
            "type": "number",
            "example": 30.0626
          },
          "longitude": {
            "type": "number",
            "example": 31.3219
          },
          "bedrooms": {
            "type": "number",
            "example": 3
          },
          "bathrooms": {
            "type": "number",
            "example": 2
          },
          "areaSize": {
            "type": "number",
            "example": 180,
            "description": "Area in square meters"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "SALE",
              "RENT"
            ],
            "example": "SALE"
          },
          "categoryId": {
            "type": "string",
            "example": "uuid-of-subcategory",
            "description": "Leaf subcategory ID"
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "city",
          "area",
          "address",
          "purpose",
          "categoryId"
        ]
      },
      "UpdatePropertyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Modern 3BR Apartment in Nasr City"
          },
          "description": {
            "type": "string",
            "example": "Spacious apartment with balcony and parking."
          },
          "price": {
            "type": "number",
            "example": 2500000
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "area": {
            "type": "string",
            "example": "Nasr City"
          },
          "address": {
            "type": "string",
            "example": "123 Abbas El Akkad Street"
          },
          "latitude": {
            "type": "number",
            "example": 30.0626
          },
          "longitude": {
            "type": "number",
            "example": 31.3219
          },
          "bedrooms": {
            "type": "number",
            "example": 3
          },
          "bathrooms": {
            "type": "number",
            "example": 2
          },
          "areaSize": {
            "type": "number",
            "example": 180,
            "description": "Area in square meters"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "SALE",
              "RENT"
            ],
            "example": "SALE"
          },
          "categoryId": {
            "type": "string",
            "example": "uuid-of-subcategory",
            "description": "Leaf subcategory ID"
          }
        }
      },
      "UploadPropertyImagesDto": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            },
            "description": "One or more property images (JPEG/PNG/WebP, max 10 MB each)"
          },
          "primaryIndex": {
            "type": "number",
            "example": 0,
            "description": "Index of uploaded file to mark as primary (0-based)"
          }
        },
        "required": [
          "images"
        ]
      },
      "UpdatePropertyImageDto": {
        "type": "object",
        "properties": {
          "isPrimary": {
            "type": "boolean",
            "example": true
          },
          "order": {
            "type": "number",
            "example": 1
          }
        }
      },
...
    }
  }
}